yt-comments/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
djangorestframework==3.14.0
idna==3.4
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
